+++ CHANGELOG.md
@@ -21,7 +21,8 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Add a build of RSS fill of the changelog (available on https://hub.openrainbow.com/#/documentation/doc/sdk/node/guides/CHANGELOGRSS.xml)
 -   Add automatic version of the SDK in jsdoc of service's classes.
 -   Add BubblesService::archiveBubble method to  close the room in one step. The other alternative is to change the status for each room users not deactivated yet. All users currently having the status 'invited' or 'accepted' will receive a message/stanza .
--   Typescript improvement
+-   Typescript improvement.
+-   Add CDD methods in TelephonyService : logon, logoff, withdrawal, wrapup.
 
 ## [1.64.2] - 2019-11-26
 -   rebundle of the 1.64.0 version with the same content.